added button and collapsing recipients, did not add for products. Can add later if needed
