+++ src/data/redux/actions/actions.js
@@ -52,6 +52,8 @@ export const UPDATE_PRODUCT_LENGTH = "UPDATE_PRODUCT_LENGTH";
 export const UPDATE_EXPIRATION_DATE = "UPDATE_EXPIRATION_DATE";
 export const UPDATE_SCHOOL = "UPDATE_SCHOOL";
 
+export const UPDATE_RECIPIENT_LIST_OPEN = "UPDATE_RECIPIENT_LIST_OPEN";
+
 export const UPDATE_PAYMENTS_SEARCH_VALUE = "UPDATE_PAYMENTS_SEARCH_VALUE";
 export const UPDATE_PAYMENTS_SUGGESTIONS = "UPDATE_PAYMENTS_SUGGESTIONS";
 export const UPDATE_PAYMENTS_SEARCH_BY = "UPDATE_PAYMENTS_SEARCH_BY";
